+++ example-app/plugin-outputs/server-route.tsx
@@ -1,28 +1,24 @@
 "poor man's use server";
 
 import { createRpcMethod as _createRpcMethod, createRpcHandler as _createRpcHandler } from "server-actions-for-next-pages/dist/server";
-import { cookies, headers } from 'server-actions-for-next-pages/headers';
-export function wrapMethod(fn) {
-  return async (...args) => {
-    try {
-      const res = await fn(...args);
-      return res;
-    } catch (error) {
-      // console.error(error);
-      throw error;
-    }
-  };
-}
-export const appServerAction = _createRpcMethod(async function appServerAction({}) {
+import fs from 'fs';
+import { revalidatePath } from 'next/cache';
+export const sendMessage = _createRpcMethod(async function sendMessage({
+  text
+}) {
   // console.log('edge cookies & headers', cookies(), headers());
-
-  return {
-    cookies: cookies().toString().slice(0, 20),
-    headers: Array.from(headers().keys()).slice(0, 2),
-    functionName: 'appRouteAction'
-  };
+  await sleep(100);
+  await fs.promises.writeFile('./optimistic.json', JSON.stringify([...JSON.parse(fs.readFileSync('./optimistic.json', 'utf-8')), {
+    text,
+    sending: false
+  }], null, 2));
+  revalidatePath('/optimistic');
+  return {};
 }, {
-  name: "appServerAction",
-  pathname: "/app-actions"
-}, typeof wrapMethod === 'function' ? wrapMethod : undefined);
-export default /*#__PURE__*/_createRpcHandler([["appServerAction", appServerAction]], false);
+  name: "sendMessage",
+  pathname: "/optimistic/actions"
+}, null);
+function sleep(ms: number) {
+  return new Promise(resolve => setTimeout(resolve, ms));
+}
+export default /*#__PURE__*/_createRpcHandler([["sendMessage", sendMessage]], false);